created chatUtils
