+++ public/css/canvas.css
@@ -848,13 +848,6 @@ button {
   margin: 30px;
 }
 
-.image-container img{
-  max-height: 360px;
-  height: 60vw;
-  border-radius: 0.75rem;
-  box-shadow: 0 19px 38px rgba(0, 0, 0, 0.3), 0 15px 12px rgba(0, 0, 0, 0.22);
-}
-
 .wrap-div-message-image{
   position: absolute;
   top: 50%;
@@ -1592,6 +1585,21 @@ div.messages.active-chat{
   border-color: blue;
   left: 20px;
   top: 10px;
+}
+
+.image-container img{
+  border-radius: 0.45rem;
+  box-shadow: 0 19px 38px rgba(0, 0, 0, 0.3), 0 15px 12px rgba(0, 0, 0, 0.22);
+}
+
+.modal-message-image-vertical{
+  max-height: 360px;
+  height: 60vw;
+}
+
+.modal-message-image-horizontal{
+  max-width: 800px;
+  width: 85vw;
 }
 
 #mic-buttons-container{
